(code review) Remove public Log API
The previous solution to allow public browsing of chat logs used a javascript-based solution that uses a public REST API. Now that this solution has been replaced, this API is no longer used. As it was never publicly announced as a 'supported' API, and maintaining it introduces overhead and complexity, this commit drops that API.
