+++ classes/WEB-INF/web.xml
@@ -5,23 +5,9 @@
          version="3.1">
 
     <servlet>
-        <servlet-name>JerseyApplicationPublic</servlet-name>
-        <servlet-class>org.glassfish.jersey.servlet.ServletContainer</servlet-class>
-        <init-param>
-            <param-name>javax.ws.rs.Application</param-name>
-            <param-value>org.jivesoftware.openfire.plugin.JerseyApplicationPublic</param-value>
-        </init-param>
-    </servlet>
-
-    <servlet>
         <servlet-name>LogsBrowserServlet</servlet-name>
         <servlet-class>org.jivesoftware.openfire.plugin.web.LogsBrowserServlet</servlet-class>
     </servlet>
-
-    <servlet-mapping>
-        <servlet-name>JerseyApplicationPublic</servlet-name>
-        <url-pattern>/api/*</url-pattern>
-    </servlet-mapping>
 
     <servlet-mapping>
         <servlet-name>LogsBrowserServlet</servlet-name>
